requirements.txt
flask
flask-mongoengine
python-dotenv
python-chess
dnspython
gunicorn
flask-socketio
flask-cors
flask-bcrypt
pyjwt
schema